Renamed namespace and cleaned up
